requirements.txt
alembic==1.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.6
typing_extensions==4.5.0
urllib3==1.26.15
